move constants.py
